# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles and their expenses
      responses:
        "200":
          description: Vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehiclesResponse"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a vehicle using a caller supplied stable ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "409":
          description: ID already exists or was deleted
  /vehicles/import:
    post:
      operationId: importVehicles
      tags: [sync]
      summary: Add missing legacy vehicles and expenses without overwriting server records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleImportInput"
      responses:
        "200":
          description: Current shared vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehiclesResponse"
  /vehicles/{vehicleId}:
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update one vehicle with optimistic conflict detection
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Vehicle not found
        "409":
          description: Version conflict
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete one vehicle and tombstone its stable ID
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "204":
          description: Vehicle deleted
        "404":
          description: Vehicle not found
  /vehicles/{vehicleId}/expenses:
    post:
      operationId: createVehicleExpense
      tags: [expenses]
      summary: Add one expense to a vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Vehicle not found
  /expenses/{expenseId}:
    patch:
      operationId: updateVehicleExpense
      tags: [expenses]
      summary: Update one expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Expense not found
        "409":
          description: Version conflict
    delete:
      operationId: deleteVehicleExpense
      tags: [expenses]
      summary: Delete one expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "204":
          description: Expense deleted
        "404":
          description: Expense not found
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, year, make, model, engine, fuel, mileage, purchasePrice, status, expenses, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          minLength: 1
        year:
          type: integer
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        engine:
          type: string
        fuel:
          type: string
        mileage:
          type: integer
          minimum: 0
        purchasePrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [active, sold]
        salePrice:
          type: ["number", "null"]
          minimum: 0
        askingPrice:
          type: ["number", "null"]
          minimum: 0
        purchaseDate:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        registration:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        soldAt:
          type: ["string", "null"]
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VehiclesResponse:
      type: array
      items:
        $ref: "#/components/schemas/Vehicle"
    VehicleInput:
      type: object
      required: [id, year, make, model, engine, fuel, mileage, purchasePrice]
      properties:
        id:
          type: string
          minLength: 1
        year:
          type: integer
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        engine:
          type: string
        fuel:
          type: string
        mileage:
          type: integer
          minimum: 0
        purchasePrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [active, sold]
        salePrice:
          type: number
          minimum: 0
        askingPrice:
          type: number
          minimum: 0
        purchaseDate:
          type: string
        vin:
          type: string
        registration:
          type: string
        location:
          type: string
        source:
          type: string
        notes:
          type: string
        soldAt:
          type: string
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseInput"
    VehicleUpdate:
      allOf:
        - $ref: "#/components/schemas/VehicleInputPartial"
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion:
              type: integer
              minimum: 1
    VehicleInputPartial:
      type: object
      properties:
        year:
          type: integer
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        engine:
          type: string
        fuel:
          type: string
        mileage:
          type: integer
          minimum: 0
        purchasePrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [active, sold]
        salePrice:
          type: ["number", "null"]
          minimum: 0
        askingPrice:
          type: ["number", "null"]
          minimum: 0
        purchaseDate:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        registration:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        soldAt:
          type: ["string", "null"]
    Expense:
      type: object
      required: [id, vehicleId, label, amount, date, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          minLength: 1
        vehicleId:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
        date:
          type: string
        category:
          type: ["string", "null"]
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [label, amount, date]
      properties:
        id:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
        date:
          type: string
        category:
          type: string
    ExpenseUpdate:
      type: object
      required: [expectedVersion]
      properties:
        label:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
        date:
          type: string
        category:
          type: ["string", "null"]
        expectedVersion:
          type: integer
          minimum: 1
    VehicleImportItem:
      allOf:
        - $ref: "#/components/schemas/VehicleInput"
        - type: object
          required: [createdAt]
          properties:
            createdAt:
              type: string
              format: date-time
            expenses:
              type: array
              items:
                $ref: "#/components/schemas/ExpenseImportItem"
    ExpenseImportItem:
      allOf:
        - $ref: "#/components/schemas/ExpenseInput"
        - type: object
          required: [id, createdAt]
          properties:
            id:
              type: string
              minLength: 1
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    VehicleImportInput:
      type: object
      required: [vehicles]
      properties:
        vehicles:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/VehicleImportItem"
...