# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  vehicleId: string;
  /** @minLength 1 */
  label: string;
  /** @minimum 0 */
  amount: number;
  date: string;
  /** @nullable */
  category?: string | null;
  /** @minimum 1 */
  version: number;
  createdAt: Date;
  updatedAt: Date;
}