# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  /** @minLength 1 */
  label?: string;
  /** @minimum 0 */
  amount?: number;
  date?: string;
  /** @nullable */
  category?: string | null;
  /** @minimum 1 */
  expectedVersion: number;
}