# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Expense } from './expense';
import type { VehicleStatus } from './vehicleStatus';

export interface Vehicle {
  /** @minLength 1 */
  id: string;
  year: number;
  /** @minLength 1 */
  make: string;
  /** @minLength 1 */
  model: string;
  engine: string;
  fuel: string;
  /** @minimum 0 */
  mileage: number;
  /** @minimum 0 */
  purchasePrice: number;
  status: VehicleStatus;
  /**
     * @minimum 0
     * @nullable
     */
  salePrice?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  askingPrice?: number | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  vin?: string | null;
  /** @nullable */
  registration?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  soldAt?: string | null;
  expenses: Expense[];
  /** @minimum 1 */
  version: number;
  createdAt: Date;
  updatedAt: Date;
}