# lib/api-zod/src/generated/types/vehicleImportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleImportItem } from './vehicleImportItem';

export interface VehicleImportInput {
  /** @maxItems 1000 */
  vehicles: VehicleImportItem[];
}