# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseInput } from './expenseInput';
import type { VehicleInputStatus } from './vehicleInputStatus';

export interface VehicleInput {
  /** @minLength 1 */
  id: string;
  year: number;
  /** @minLength 1 */
  make: string;
  /** @minLength 1 */
  model: string;
  engine: string;
  fuel: string;
  /** @minimum 0 */
  mileage: number;
  /** @minimum 0 */
  purchasePrice: number;
  status?: VehicleInputStatus;
  /** @minimum 0 */
  salePrice?: number;
  /** @minimum 0 */
  askingPrice?: number;
  purchaseDate?: string;
  vin?: string;
  registration?: string;
  location?: string;
  source?: string;
  notes?: string;
  soldAt?: string;
  expenses?: ExpenseInput[];
}