# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleInputPartial } from './vehicleInputPartial';

export type VehicleUpdate = VehicleInputPartial & {
  /** @minimum 1 */
  expectedVersion: number;
};